app+ui: maintain the connection list when paused
Previously the connections would get overwritten when updating. This was
most apparent when the program was in the pause state, and trying to toggle
through the different tabs would not change the connections.

This separates out the connections and the filtering process, so that the
connections are all preserved regardless of state/tab selection.