os: add basic support for Linux
While this works, there could be improvements. Particularly because it
has to traverse the all_processes call each time when we technically
already know the pid.

There's probably a way to do something like procfs::by_pid(pid) so we
don't have to go through the whole pid list, but this will work for now.